Hackers multitool/requirements.txt
colorama
requests
python-whois
ipwhois
termcolor 